Fixed some 'longdiv' display issues in IE8.
The 'longdiv' display was wonky because I had not set a width on the
problemContainer.

Made the 'longdiv' hints block presentable.

The demo "popup" window is positioned to the right of the table.problem
element, not the problemContainer.

Changed the numpad submit symbol to equals sign so that it appears in
all browsers.